requirements.txt
tensorflow>=2.13
numpy
Pillow
opencv-python
scikit-learn
kagglehub
matplotlib
